fix(agdi): update APBank when access AP register
When using PDSC Debug, the APBank may be updated directly and the
associated AP access function in SWD/JTAG does not perceive the change.
So we record the value of the last APBank in these functions and update
it when needed to ensure that the AP registers accessed are as
expected.