fix(agdi): ensure DSM_SuspendMonitor() can enter the suspend state
In DSM_SuspendMonitor(), the suspendCount may be greater than 0, but
it is not in a suspended state.

This is because in the DSMonitor_Thread(), the ReadD32() will
actively perform suspend and resume to avoid conflicts. When reaching
this point, it may just be between suspend and resume in ReadD32(),
meaning suspendCount > 0, but inSuspension == 0.

At this time, the debugger is still active. It is necessary to ensure
that after the execution of ReadD32(), it can enter the suspend state
again.